Final 1.0
projeto para rodar no servidor

+++ users/models.py
@@ -3,7 +3,7 @@ from django.db import models
 
 
 class User(AbstractUser):
-    espaco = models.IntegerField(blank=True,default=0)
+    espaco = models.FloatField(blank=True,default=0)
     premium = models.CharField(max_length=5, default="0")
 
     def __str__(self):
